Split the class declarations and implementations.
