+++ Boss.h
@@ -1,6 +1,5 @@
-#include "Arduino.h"
-
-#include <stdint.h> // uint8_t type variables
+#ifndef BOSS_H
+#define BOSS_H
 
 class Boss {
     public:
@@ -15,29 +14,4 @@ class Boss {
         bool _alive;
 };
 
-void Boss::Spawn() {
-    _pos = 800;
-    _lives = 3;
-    _alive = 1;
-}
-
-void Boss::Hit() {
-    _lives --;
-    if (_lives == 0) {
-        Kill();
-        return;
-    }
-    if (_lives == 2) {
-        _pos = 200;
-    } else if(_lives == 1) {
-        _pos = 600;
-    }
-}
-
-bool Boss::Alive() {
-    return _alive;
-}
-
-void Boss::Kill() {
-    _alive = 0;
-}
+#endif /* BOSS_H */
