+++ Conveyor.h
@@ -1,4 +1,5 @@
-#include "Arduino.h"
+#ifndef CONVEYOR_H
+#define CONVEYOR_H
 
 class Conveyor {
     public:
@@ -10,13 +11,4 @@ class Conveyor {
         bool _alive;
 };
 
-void Conveyor::Spawn(int startPoint, int endPoint, int speed){
-    _startPoint = startPoint;
-    _endPoint = endPoint;
-    _speed = speed;  // must be less than +/- max player speed
-    _alive = true;
-}
-
-void Conveyor::Kill(){
-    _alive = false;
-}
+#endif /* CONVEYOR_H */
