+++ Lava.h
@@ -1,4 +1,5 @@
-#include "Arduino.h"
+#ifndef LAVA_H
+#define LAVA_H
 
 /*
 startPoint: 0 to 1000
@@ -25,21 +26,4 @@ class Lava {
         bool _alive;
 };
 
-void Lava::Spawn(int left, int right, int ontime, int offtime, int offset, bool state) {
-    _left = left;
-    _right = right;
-    _ontime = ontime;
-    _offtime = offtime;
-    _offset = offset;
-    _alive = 1;
-    _lastOn = millis()-offset;
-    _state = state;
-}
-
-void Lava::Kill() {
-    _alive = 0;
-}
-
-bool Lava::Alive() {
-    return _alive;
-}
+#endif /* LAVA_H */
