+++ iSin.h
@@ -1,4 +1,5 @@
-#include "Arduino.h"
+#ifndef ISIN_H
+#define ISIN_H
 
 class iSin {
     public:
@@ -17,22 +18,4 @@ class iSin {
         };
 };
 
-int iSin::convert(long x)
-{
-    boolean pos = true;  // positive - keeps an eye on the sign.
-    if (x < 0)
-    {
-        x = -x;
-        pos = !pos;
-    }
-    if (x >= 360) x %= 360;
-    if (x > 180)
-    {
-        x -= 180;
-        pos = !pos;
-    }
-    if (x > 90) x = 180 - x;
-    if (pos) return pgm_read_byte_near(isinTable8[x])/2 ;
-    return -isinTable8[x]/2 ;
-}
-
+#endif /* ISIN_H */
